+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/BaneOfSSSS.java
@@ -7,15 +7,17 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
-import net.minecraft.world.entity.*;
-import net.minecraft.world.entity.monster.Creeper;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
 public class BaneOfSSSS extends BonusDamageEnchantment {
     public static final TagKey<EntityType<?>> AFFECTED_BY_BANE_OF_SSSSS = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(IguanaTweaksReborn.MOD_ID, "enchantments/bane_of_sssss"));
     public BaneOfSSSS() {
-        super(Rarity.UNCOMMON, EnchantmentsFeature.ITR_WEAPONS, new EquipmentSlot[]{ EquipmentSlot.MAINHAND });
+        super(Rarity.UNCOMMON, EnchantmentsFeature.ITR_WEAPONS_CATEGORY, new EquipmentSlot[]{ EquipmentSlot.MAINHAND });
     }
 
     @Override
@@ -23,7 +25,7 @@ public class BaneOfSSSS extends BonusDamageEnchantment {
         if (!(target instanceof LivingEntity livingentity))
             return;
 
-        if (lvl > 0 && (livingentity.getMobType() == MobType.ARTHROPOD || livingentity instanceof Creeper)) {
+        if (lvl > 0 && livingentity.getType().is(AFFECTED_BY_BANE_OF_SSSSS)) {
             int i = 20 + attacker.getRandom().nextInt(10 * lvl);
             livingentity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SLOWDOWN, i, 3));
         }
@@ -37,7 +39,7 @@ public class BaneOfSSSS extends BonusDamageEnchantment {
     }
 
     @Override
-    public float getDamageBonus(ItemStack stack, int lvl) {
-        return 1.5f * lvl * this.getDamageBonusRatio(stack);
+    public boolean isAffectedByEnchantment(LivingEntity target) {
+        return target.getType().is(AFFECTED_BY_BANE_OF_SSSSS);
     }
 }